add AbstractRWLogger abstract class
